windows_app/requirements.txt
pywebview==4.4.1
pythonnet==3.0.3